move sidechain-specific code out of selection crate
