+++ toolkit/primitives/selection/Cargo.toml
@@ -10,8 +10,6 @@ license = "Apache-2.0"
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-sp-std = { workspace = true }
-sp-core = { workspace = true }
 rand_chacha = { workspace = true }
 rand = { workspace = true }
 parity-scale-codec = { workspace = true }
@@ -19,8 +17,6 @@ scale-info = { workspace = true }
 sidechain-domain = { workspace = true }
 
 [dev-dependencies]
-sp-io = { workspace = true }
-sp-runtime = { workspace = true }
 quickcheck = { workspace = true }
 quickcheck_macros = { workspace = true }
 
@@ -28,8 +24,6 @@ quickcheck_macros = { workspace = true }
 default = ["std"]
 std = [
 	"sidechain-domain/std",
-    "sp-std/std",
-    "sp-core/std",
 	"parity-scale-codec/std",
 	"scale-info/std",
     "rand/std"
